Add depthwise conv2d for Theano and CNTK
