+++ tests/keras/backend/backend_test.py
@@ -844,27 +844,31 @@ class TestBackend(object):
             with pytest.raises(ValueError):
                 k.conv3d(k.variable(xval), k.variable(kernel_val), data_format='channels_middle')
 
-    @pytest.mark.parametrize('k', [KTF], ids=['TensorFlow'])
-    def test_depthwise_conv_2d(self, k):
-        for data_format in ['channels_first', 'channels_last']:
-            x_shape = (4, 4)
-            if data_format == 'channels_first':
-                input_shape = (2, 3) + x_shape
-            elif data_format == 'channels_last':
-                input_shape = (2,) + x_shape + (3,)
-            kernel_shape = (3, 3, 3, 2)
+    def test_depthwise_conv_2d(self):
+        # TF kernel shape: (rows, cols, input_depth, depth_multiplier)
+        # channels_first input shape: (n, input_depth, rows, cols)
+        for input_shape in [(2, 3, 4, 5), (2, 3, 5, 6)]:
+            for kernel_shape in [(2, 2, 3, 4), (4, 3, 3, 4)]:
+                check_two_tensor_operation('depthwise_conv2d',
+                                           input_shape, kernel_shape,
+                                           BACKENDS, cntk_dynamicity=True,
+                                           data_format='channels_first')
 
-            x_val = np.ones(input_shape)
-            kernel_val = np.arange(np.prod(kernel_shape)).reshape(kernel_shape)
-            z = k.eval(k.depthwise_conv2d(k.variable(x_val), k.variable(kernel_val),
-                                          data_format=data_format))
+        input_shape = (1, 6, 5, 3)
+        kernel_shape = (3, 3, 3, 2)
+        check_two_tensor_operation('depthwise_conv2d',
+                                   input_shape, kernel_shape,
+                                   BACKENDS, cntk_dynamicity=True,
+                                   data_format='channels_last')
 
-            for z_i in np.split(z, 6, axis=1 if data_format == 'channels_first' else -1):
-                assert_allclose(z_i, z_i[0] * np.ones_like(z_i))
-
+        xval = np.random.random(input_shape)
+        kernel_val = np.random.random(kernel_shape) - 0.5
         # Test invalid use cases
-        with pytest.raises(ValueError):
-            k.depthwise_conv2d(k.variable(x_val), k.variable(kernel_val), data_format='channels_middle')
+        for k in BACKENDS:
+            with pytest.raises(ValueError):
+                k.depthwise_conv2d(k.variable(xval),
+                                   k.variable(kernel_val),
+                                   data_format='channels_middle')
 
     def test_pool2d(self):
         check_single_tensor_operation('pool2d', (5, 10, 12, 3),
